Removed constructor in Ticket - using lombok moved validation to service class
