+++ src/main/java/com/vmtapp/enterprise/service/TicketServiceStub.java
@@ -5,6 +5,8 @@ import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @Component
 public class TicketServiceStub implements ITicketService {
@@ -13,13 +15,31 @@ public class TicketServiceStub implements ITicketService {
     public ArrayList<Ticket> fetchTicketsByAssignee(String assignee) {
         ArrayList<Ticket> ticketsToReturn = new ArrayList<Ticket>();
 
-        Ticket ticket = new Ticket("Larry","Fine","larry@test.com");
+        Ticket ticket = new Ticket();
+        ticket.setFirstname("Larry");
+        ticket.setLastname("Fine");
+        ticket.setEmail("larry@test.com");
         ticketsToReturn.add(ticket);
-        ticket = new Ticket("Moe","Howard","moe@test.com");
+        ticket.setFirstname("Moe");
+        ticket.setLastname("Howard");
+        ticket.setEmail("moe@test.com");
         ticketsToReturn.add(ticket);
-        ticket = new Ticket("Shemp","Howard","shemp@test.com");
+        ticket.setFirstname("Shemp");
+        ticket.setLastname("Howard");
+        ticket.setEmail("shemp@test.com");
         ticketsToReturn.add(ticket);
 
         return ticketsToReturn;
+    }
+
+    private boolean emailValidation(String email)
+    {
+        String emailRegEx = "^[A-Za-z0-9+_.-]+@(.+)$";
+
+        Pattern pattern = Pattern.compile(emailRegEx);
+
+        Matcher matcher = pattern.matcher(email);
+
+        return matcher.find();
     }
 }
